--- test/app.local.js
@@ -1,6 +1,0 @@
-'use strict'
-const app = require('./app')
-const port = process.env.PORT || 5000
-app.listen(port, () =>
-  console.log(`Server is listening on port ${port}.`)
-)
