+++ test/netlify.toml
@@ -1,0 +1,3 @@
+[build]
+    functions = "dist"
+
